refactor: views -> screens (folder renamed)
