+++ app/src/main/java/uk/nktnet/webviewkiosk/ui/screens/SettingsAboutScreen.kt
@@ -1,4 +1,4 @@
-package uk.nktnet.webviewkiosk.ui.view
+package uk.nktnet.webviewkiosk.ui.screens
 
 import android.content.pm.PackageManager
 import android.os.Build
