+++ app/src/main/java/uk/nktnet/webviewkiosk/ui/screens/SettingsWebEngineScreen.kt
@@ -1,4 +1,4 @@
-package uk.nktnet.webviewkiosk.ui.view
+package uk.nktnet.webviewkiosk.ui.screens
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
@@ -9,11 +9,10 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import uk.nktnet.webviewkiosk.ui.components.setting.SettingDivider
 import uk.nktnet.webviewkiosk.ui.components.setting.SettingLabel
-import uk.nktnet.webviewkiosk.ui.components.setting.fielditems.weblifecycle.ResetOnInactivitySecondsSetting
-import uk.nktnet.webviewkiosk.ui.components.setting.fielditems.weblifecycle.ResetOnLaunchSetting
+import uk.nktnet.webviewkiosk.ui.components.setting.fielditems.webengine.*
 
 @Composable
-fun SettingsWebLifecycleScreen(navController: NavController) {
+fun SettingsWebEngineScreen(navController: NavController) {
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -21,11 +20,20 @@ fun SettingsWebLifecycleScreen(navController: NavController) {
             .windowInsetsPadding(WindowInsets.safeContent)
             .verticalScroll(rememberScrollState())
     ) {
-        SettingLabel(navController = navController, label = "Web Lifecycle")
+        SettingLabel(navController = navController, label = "Web Engine")
 
         SettingDivider()
 
-        ResetOnLaunchSetting()
-        ResetOnInactivitySecondsSetting()
+        EnableJavaScriptSetting()
+        EnableDomStorageSetting()
+        AcceptCookiesSetting()
+        AcceptThirdPartyCookiesSetting()
+        CacheModeSetting()
+        UserAgentSetting()
+        LayoutAlgorithmSetting()
+        UseWideViewPortSetting()
+        LoadWithOverviewModeSetting()
+        EnableZoomSetting()
+        DisplayZoomControlsSetting()
     }
 }
